refactor(pages): chuẩn hóa tên file và cấu trúc thư mục - Đổi tên các file page thành: AboutUsPage, HomePage, NotFoundPage,   ShopPage, ShopDetailPage, AuthPage - Di chuyển các component vào thư mục components tương ứng - Cập nhật lại import và routes cho phù hợp
