--- src/pages/Home/Home.tsx
@@ -1,13 +1,0 @@
-import AboutUs from "./AboutUs";
-import Hero from "./Hero";
-
-function Home() {
-  return (
-    <>
-      <Hero />
-      <AboutUs />
-    </>
-  );
-}
-
-export default Home;
