+++ src/pages/Home/HomePage.tsx
@@ -1,0 +1,13 @@
+import AboutUs from "./components/AboutUs";
+import Hero from "./components/Hero";
+
+function HomePage() {
+  return (
+    <>
+      <Hero />
+      <AboutUs />
+    </>
+  );
+}
+
+export default HomePage;
