+++ src/pages/Home/components/AboutUs.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography } from "@mui/material";
 import { motion } from "framer-motion";
+import MyButton from "../../../components/common/Button";
 import ImageAboutUs from "../../assets/images/image_about_us_home.png";
-import MyButton from "../../components/common/Button";
 
 function AboutUs() {
   return (
