+++ src/routes.tsx
@@ -1,22 +1,22 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
 import RedirectIfAuth from "./components/common/RedirectIfAuth";
-import AuthPage from "./pages/Auth/Auth";
-import Home from "./pages/Home/Home";
-import NotFound from "./pages/NotFound/NotFound";
-import ShopDetail from "./pages/ShopDetail/ShopDetail";
-import OurShop from "./pages/Shop/Shop";
-import AboutUs from "./pages/AboutUs/AboutUS";
+import AuthPage from "./pages/Auth/AuthPage";
+import HomePage from "./pages/Home/HomePage";
+import NotFoundPage from "./pages/NotFound/NotFoundPage";
+import ShopDetailPage from "./pages/ShopDetail/ShopDetailPage";
+import OurShopPage from "./pages/Shop/ShopPage";
+import AboutUsPage from "./pages/AboutUs/AboutUsPage";
 const router = createBrowserRouter([
   {
     path: "/",
     element: <App />,
     children: [
-      { path: "", element: <Home /> },
-      { path: "shopdetail", element: <ShopDetail /> },
-      { path: "about", element: <AboutUs /> },
-      { path: "ourshop", element: <OurShop /> },
-      { path: "*", element: <NotFound /> },
+      { path: "", element: <HomePage /> },
+      { path: "shopdetail", element: <ShopDetailPage /> },
+      { path: "about", element: <AboutUsPage /> },
+      { path: "ourshop", element: <OurShopPage /> },
+      { path: "*", element: <NotFoundPage /> },
     ],
   },
   {
